+++ rust-maven-plugin/src/main/java/io/questdb/maven/rust/Shlex.java
@@ -22,25 +22,37 @@
  *
  ******************************************************************************/
 
-package io.questdb.rust.maven.example;
+package io.questdb.maven.rust;
 
-import io.questdb.jar.jni.JarJniLoader;
+import java.util.List;
+import java.util.regex.Pattern;
 
-public class Main {
-    public static native String reversedString(String str);
+public interface Shlex {
 
-    static {
-        JarJniLoader.loadLib(
-                Main.class,
-
-                // A platform-specific path is automatically suffixed to path below.
-                "/io/questdb/rust/maven/example/libs",
-
-                // The "lib" prefix and ".so|.dynlib|.dll" suffix are added automatically as needed.
-                "str_reverse");
+    /**
+     * Escape a string for use in a shell command.
+     *
+     * @param s The string to escape.
+     * @return The escaped string.
+     * @see <a href="https://docs.python.org/3/library/shlex.html#shlex.quote">shlex.quote</a>
+     */
+    static String quote(String s) {
+        if (s.isEmpty())
+            return "''";
+        Pattern unsafe = Pattern.compile("[^\\w@%+=:,./-]");
+        if (unsafe.matcher(s).find())
+            return "'" + s.replace("'", "'\"'\"'") + "'";
+        else
+            return s;
     }
 
-    public static void main(String[] args) {
-        System.out.println(reversedString("Hello World!"));
+    static String quote(List<String> args) {
+        StringBuilder sb = new StringBuilder();
+        for (String arg : args) {
+            if (sb.length() > 0)
+                sb.append(' ');
+            sb.append(quote(arg));
+        }
+        return sb.toString();
     }
 }
